Fix "uncontroversial" clippy lints and start making them errors in TravisCI
Which lints are clearly ones we should fix (and fail on in the future) is
subjective, but this is a first cut at making things better, so that we can
be more idiomatic, consistent and performant going forward and clippy can
help let us know when we stray off the path.

Deciding what to do about the remaining lints in the UNEXAMINED_LINTS
category in the Makefile will be an exercise for another PR.

+++ components/hab/src/command/supportbundle.rs
@@ -70,16 +70,13 @@ pub fn start(ui: &mut UI) -> Result<()> {
             Status::Adding,
             format!("files from {}", &sup_root.display()),
         )?;
-        match tar.append_dir_all(format!("hab{}sup", MAIN_SEPARATOR), &sup_root) {
-            Err(why) => {
-                ui.fatal(format!(
-                    "Failed to add all files into the tarball: {:?}",
-                    why.description()
-                ))?;
-                fs::remove_file(&tarball_name)?;
-                process::exit(1)
-            }
-            Ok(_) => {}
+        if let Err(why) = tar.append_dir_all(format!("hab{}sup", MAIN_SEPARATOR), &sup_root) {
+            ui.fatal(format!(
+                "Failed to add all files into the tarball: {:?}",
+                why.description()
+            ))?;
+            fs::remove_file(&tarball_name)?;
+            process::exit(1);
         }
     } else {
         ui.fatal(format!(
